requirements.txt
selenium~=4.32.0
webdriver-manager~=4.0.2
python-dotenv~=1.1.0